requirements.txt
fastapi
langchain-google-genai
langchain-community
langchain-pinecone
langchain-huggingface
sentence-transformers
pypdf
streamlit
uvicorn
python-dotenv
